refactor: files outside src directory
