+++ client/tests/unit/components/reusables/LoadingSpinner/LoadingSpinner.unit.test.tsx
@@ -4,19 +4,15 @@ import { describe, it, expect } from "vitest";
 import { LoadingSpinner } from "../../../../../src/components/reusables/LoadingSpinner/LoadingSpinner";
 
 describe("LoadingSpinner component", () => {
-  it("should render the loading image for the spinner", () => {
+  it("should render the loading image & content for the spinner", () => {
     render(<LoadingSpinner content="Loading..." />);
 
+    // user should see the loading spinner image
     const image = screen.getByRole("img", { name: /loading./i });
-
     expect(image).toBeInTheDocument();
-  });
 
-  it("should render the content for loading spinner", () => {
-    render(<LoadingSpinner content="Loading..." />);
-
+    // user should see the loading spinner content
     const content = screen.getByText(/loading.../i);
-
     expect(content).toBeInTheDocument();
   });
 });
